# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure MFA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  id: string;
  /** Event kind, e.g. login_success, login_failed, register, enroll_face */
  kind: string;
  /** Method used (mpin, face, biometric, otp) */
  method?: string | null;
  success: boolean;
  createdAt: Date;
  ipAddress?: string | null;
}