# lib/api-zod/src/generated/types/loginChallengeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure MFA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LoginChallengeResponseUserHint } from "./loginChallengeResponseUserHint";

export interface LoginChallengeResponse {
  challengeToken: string;
  availableFactors: string[];
  userHint: LoginChallengeResponseUserHint;
}