# lib/api-zod/src/generated/types/loginWebauthnOptionsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Secure MFA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginWebauthnOptionsBody {
  challengeToken: string;
}